t2131: refactor(misc-scripts): decompose voice-bridge/normalise-markdown/tabby-profile-sync
Extract high-complexity Python files into focused helper modules to
eliminate qlty file-complexity smells (voice-bridge: 95, normalise-markdown: 80,
tabby-profile-sync: 76).

normalise-markdown.py (80→0):
  - normalise_markdown_headings.py: heading detection + hierarchy helpers
  - normalise_markdown_email.py: email section detection handlers

tabby-profile-sync.py (76→0):
  - tabby_colour_utils.py: DARK_SCHEMES data + colour utility functions
  - tabby_yaml_helpers.py: YAML parsing and profile insertion helpers

voice-bridge.py (95→0):
  - voice_stt.py: SileroVAD + STT engines (Whisper MLX, Faster Whisper, macOS)
  - voice_tts.py: TTS engines (Edge TTS, macOS say, Facebook MMS)
  - voice_llm.py: OpenCodeBridge (LLM query layer)
  - voice_bridge_core.py: VoiceBridge main loop
  - voice_bridge_cli.py: CLI parsing, factory functions

All entry points (shell scripts call these files by name) preserved.
Existing tests pass. Verified with qlty smells.

Fixes #19227
